v6: re-export `ExecuteButton` and make it the one run button
Extract the run control out of `top-bar` into its own `ExecuteButton`
component and export it from `@graphiql/react` again, so a custom toolbar
can compose it without taking the whole `TopBar`. The top bar renders this
same component instead of its own copy.

The button stops a request again while one is in flight, as it did in v5:
it becomes a Stop button while fetching or while a subscription is active,
and stays enabled in that state even when a fresh run would be blocked.
The picker is hidden while running.

`TopBarView` takes the control through a `runButton` slot, so it stays
presentational and its tests keep driving it by props. It keeps
`runDisabledReason`, which it still needs for the method toggle's
attention state.

CSS classes move from `graphiql-top-bar-run*` to `graphiql-execute-button*`
now that the control is not top bar specific.
