+++ packages/graphiql-react/src/components/execute-button/index.css
@@ -1,0 +1,139 @@
+/* The split Run control: a primary run half and, when the document has more
+   than one operation, a caret half opening the operation picker. Both halves
+   render their own green fill so `:hover`/`:active` on either (which also
+   match this wrapper, per normal CSS bubbling) can brighten/darken the whole
+   pill as one piece. */
+.graphiql-execute-button {
+  display: inline-flex;
+  align-items: center;
+  border-radius: var(--radius-md);
+  color: #fff;
+  font-family: var(--font-family);
+  font-size: var(--font-size-body);
+  font-weight: var(--font-weight-medium);
+  white-space: nowrap;
+  flex-shrink: 0;
+}
+
+.graphiql-execute-button:hover {
+  filter: brightness(1.1);
+}
+
+.graphiql-execute-button:active {
+  filter: brightness(0.92);
+}
+
+.graphiql-execute-button-primary,
+.graphiql-execute-button-caret {
+  display: inline-flex;
+  align-items: center;
+  height: 28px;
+  background: oklch(var(--btn-primary));
+  border: 1px solid oklch(var(--btn-primary-border));
+  color: inherit;
+  font: inherit;
+  cursor: pointer;
+}
+
+/* The caret is an `UnStyledButton`, whose `all: unset` resets `box-sizing` to
+   `content-box` — its border would add to the 28px height and make it stand a
+   pixel taller than the primary half. Scope under `.graphiql-execute-button` so
+   this wins on specificity, not load order. */
+.graphiql-execute-button .graphiql-execute-button-caret {
+  box-sizing: border-box;
+}
+
+.graphiql-execute-button-primary {
+  gap: var(--px-8);
+  padding: 0 var(--px-10);
+  border-radius: var(--radius-md) 0 0 var(--radius-md);
+}
+
+/* No caret to pair with (a single operation, an override pinning it, or a run
+   in flight) — the primary half is the whole pill. */
+.graphiql-execute-button-primary--solo {
+  border-radius: var(--radius-md);
+}
+
+.graphiql-execute-button-caret {
+  padding: 0 var(--px-8);
+  border-left: none;
+  border-radius: 0 var(--radius-md) var(--radius-md) 0;
+}
+
+.graphiql-execute-button-caret-icon {
+  width: 8px;
+  height: auto;
+  flex-shrink: 0;
+}
+
+.graphiql-execute-button-icon {
+  width: auto;
+  height: 11px;
+  flex-shrink: 0;
+}
+
+.graphiql-execute-button-sep {
+  width: 1px;
+  height: 14px;
+  background: rgb(255 255 255 / 0.28);
+  flex-shrink: 0;
+}
+
+/* The shortcut keys sit on the green fill here, not the dark bar. Darken them
+   into subtle inset chips rather than the default dark boxes; a light (white)
+   overlay would lighten the green and drop the white glyphs below AA contrast. */
+.graphiql-execute-button .graphiql-keycap {
+  border-color: rgb(255 255 255 / 0.2);
+  background: rgb(0 0 0 / 0.18);
+  color: #fff;
+}
+
+.graphiql-execute-button-primary:disabled {
+  opacity: 0.6;
+  cursor: not-allowed;
+  /* Let pointer events fall through to the tooltip-target wrapper so the
+     "switch to POST" tooltip can open over the disabled button. */
+  pointer-events: none;
+}
+
+.graphiql-execute-button-caret:disabled {
+  opacity: 0.6;
+  cursor: not-allowed;
+}
+
+/* Scoped under `.graphiql-execute-button` so this outranks the global
+   `.graphiql-container :focus-visible` ring on specificity, not load order —
+   an inset ring (rather than the global's outward one) keeps the two halves'
+   rings from overlapping each other. */
+.graphiql-execute-button .graphiql-execute-button-primary:focus-visible,
+.graphiql-execute-button .graphiql-execute-button-caret:focus-visible {
+  outline: 2px solid oklch(var(--accent-blue));
+  outline-offset: -2px;
+}
+
+.graphiql-execute-button-tooltip-target {
+  display: inline-flex;
+  flex-shrink: 0;
+  cursor: not-allowed;
+}
+
+.graphiql-execute-button-menu-item {
+  display: inline-flex;
+  align-items: center;
+  gap: var(--px-6);
+}
+
+/* Every item reserves the dot's gutter so names stay left-aligned whether or
+   not the item is active; only the active item paints its dot. */
+.graphiql-execute-button-menu-item::before {
+  content: '';
+  width: 6px;
+  height: 6px;
+  border-radius: 50%;
+  flex-shrink: 0;
+}
+
+.graphiql-execute-button-menu-item--active::before {
+  background: oklch(var(--accent-green));
+}
