+++ packages/graphiql-react/src/components/index.ts
@@ -1,3 +1,4 @@
+export { ExecuteButton } from './execute-button';
 export { ToolbarButton } from './toolbar-button';
 export { ToolbarMenu } from './toolbar-menu';
