+++ packages/graphiql-react/src/components/top-bar/index.tsx
@@ -4,12 +4,10 @@
 
 import type { FC, ReactNode } from 'react';
 import type { HttpMethod } from '@graphiql/toolkit';
-import type { OperationDefinitionNode } from 'graphql';
 import { useGraphiQL, useGraphiQLActions } from '../provider';
-import { KeycapHint, MODIFIER } from '../keycap-hint';
 import { Tooltip } from '../tooltip';
-import { DropdownMenu } from '../dropdown-menu';
-import { GraphQLLogoIcon, PlayIcon, ChevronDownIcon } from '../../icons';
+import { ExecuteButton } from '../execute-button';
+import { GraphQLLogoIcon } from '../../icons';
 import { clsx } from 'clsx';
 import {
   getRunBlockReason,
@@ -28,15 +26,9 @@ export type TopBarProps = {
 };
 
 export const TopBar: FC<TopBarProps> = ({ version, brand }) => {
-  const { run, setTransportMethod, setOperationName } = useGraphiQLActions();
-  const isFetching = useGraphiQL(state => state.isFetching);
+  const { setTransportMethod } = useGraphiQLActions();
   const transport = useGraphiQL(state => state.transport);
   const transportMethod = useGraphiQL(state => state.transportMethod);
-  const operations = useGraphiQL(state => state.operations);
-  const operationName = useGraphiQL(state => state.operationName);
-  const overrideOperationName = useGraphiQL(
-    state => state.overrideOperationName,
-  );
   const runDisabledReason = useGraphiQL(state =>
     getRunBlockReason(
       state.transportMethod,
@@ -52,18 +44,12 @@ export const TopBar: FC<TopBarProps> = ({ version, brand }) => {
     <TopBarView
       version={version}
       brand={brand}
-      isFetching={isFetching}
       url={url}
       method={method}
       supportedMethods={supportedMethods}
       runDisabledReason={runDisabledReason}
-      operations={operations}
-      operationName={operationName}
-      overrideOperationName={overrideOperationName}
-      transportMethod={transportMethod}
-      onRun={run}
+      runButton={<ExecuteButton />}
       onSetMethod={setTransportMethod}
-      onSetOperationName={setOperationName}
     />
   );
 };
@@ -71,38 +57,28 @@ export const TopBar: FC<TopBarProps> = ({ version, brand }) => {
 export type TopBarViewProps = {
   version?: string;
   brand?: ReactNode;
-  isFetching: boolean;
   url: string;
   method: HttpMethod;
   supportedMethods: HttpMethod[];
-  /** Non-null when Run is blocked; the string is the reason shown in a tooltip. */
+  /**
+   * Non-null when a run is blocked; drives the method toggle's attention state.
+   * The reason itself is surfaced by the run control.
+   */
   runDisabledReason?: string | null;
-  /** The document's operations. A caret + picker only appears for more than one. */
-  operations?: OperationDefinitionNode[];
-  operationName?: string | null;
-  /** When set, an external caller has pinned the operation; the picker is hidden. */
-  overrideOperationName?: string | null;
-  transportMethod?: HttpMethod | null;
-  onRun: () => void;
+  /** The run control, rendered at the end of the bar. */
+  runButton?: ReactNode;
   onSetMethod: (method: HttpMethod) => void;
-  onSetOperationName?: (operationName: string) => void;
 };
 
 export const TopBarView: FC<TopBarViewProps> = ({
   version,
   brand,
-  isFetching,
   url,
   method,
   supportedMethods,
   runDisabledReason = null,
-  operations = [],
-  operationName = null,
-  overrideOperationName = null,
-  transportMethod = null,
-  onRun,
+  runButton,
   onSetMethod,
-  onSetOperationName,
 }) => {
   const canSwitch = supportedMethods.length > 1;
   const isBlocked = runDisabledReason !== null;
@@ -115,101 +91,6 @@ export const TopBarView: FC<TopBarViewProps> = ({
     ] ?? method;
   const switchTarget =
     isBlocked && supportedMethods.includes('POST') ? 'POST' : nextMethod;
-
-  // A picker only makes sense when there's a choice to make, and only when
-  // nothing outside the editor has already pinned the operation to run.
-  const hasOptions =
-    operations.length > 1 && typeof overrideOperationName !== 'string';
-  const activeOperation = resolveActiveOperation(operations, operationName);
-
-  const selectOperation = (selectedOperationName: string | undefined) => {
-    if (selectedOperationName && selectedOperationName !== operationName) {
-      onSetOperationName?.(selectedOperationName);
-    }
-    onRun();
-  };
-
-  const primaryButton = (
-    <button
-      type="button"
-      className={clsx(
-        'graphiql-top-bar-run-primary',
-        !hasOptions && 'graphiql-top-bar-run-primary--solo',
-      )}
-      onClick={onRun}
-      disabled={isFetching || isBlocked}
-      aria-label="Run query"
-    >
-      <PlayIcon className="graphiql-top-bar-run-icon" aria-hidden="true" />
-      <span className="graphiql-top-bar-run-label">Run</span>
-      <span className="graphiql-top-bar-run-sep" aria-hidden="true" />
-      <KeycapHint
-        keys={[MODIFIER.Meta, MODIFIER.Enter]}
-        ariaLabel="Run query shortcut"
-      />
-    </button>
-  );
-
-  const runButton = (
-    <div className="graphiql-top-bar-run">
-      {isBlocked ? (
-        <Tooltip label={runDisabledReason}>
-          {/* A native disabled button emits no pointer/focus events, so Radix
-              would never open the tooltip. Wrap it in a focusable span that
-              receives the events instead. */}
-          <span className="graphiql-top-bar-run-tooltip-target" tabIndex={0}>
-            {primaryButton}
-          </span>
-        </Tooltip>
-      ) : (
-        primaryButton
-      )}
-
-      {hasOptions && (
-        <>
-          <span className="graphiql-top-bar-run-sep" aria-hidden="true" />
-          <DropdownMenu>
-            <DropdownMenu.Button
-              type="button"
-              className="graphiql-top-bar-run-caret"
-              disabled={isFetching}
-              aria-label="Choose operation to run"
-            >
-              <ChevronDownIcon
-                className="graphiql-top-bar-run-caret-icon"
-                aria-hidden="true"
-              />
-            </DropdownMenu.Button>
-            <DropdownMenu.Content align="end">
-              {operations.map((operation, i) => {
-                const opName = operation.name?.value;
-                const label = opName ?? `<Unnamed ${operation.operation}>`;
-                const isActive = operation === activeOperation;
-                return (
-                  <DropdownMenu.Item
-                    key={`${label}-${i}`}
-                    disabled={
-                      getRunBlockReason(transportMethod, operation) !== null
-                    }
-                    onSelect={() => selectOperation(opName)}
-                  >
-                    <span
-                      className={clsx(
-                        'graphiql-top-bar-run-menu-item',
-                        isActive && 'graphiql-top-bar-run-menu-item--active',
-                      )}
-                    >
-                      {label}
-                    </span>
-                  </DropdownMenu.Item>
-                );
-              })}
-            </DropdownMenu.Content>
-          </DropdownMenu>
-        </>
-      )}
-    </div>
-  );
 
   return (
     <header className="graphiql-top-bar" role="banner">
