+++ packages/graphiql-react/src/components/top-bar/top-bar.test.tsx
@@ -5,22 +5,13 @@ import type { HttpMethod } from '@graphiql/toolkit';
 import { describe, it, expect, vi } from 'vitest';
 import { render, screen } from '@testing-library/react';
 import userEvent from '@testing-library/user-event';
-import { parse, OperationDefinitionNode } from 'graphql';
 import { TopBarView } from './';
 import { Tooltip } from '../tooltip';
 
-function opsOf(source: string): OperationDefinitionNode[] {
-  return parse(source).definitions.filter(
-    (d): d is OperationDefinitionNode => d.kind === 'OperationDefinition',
-  );
-}
-
 const DEFAULTS = {
-  isFetching: false,
   url: 'https://api.example.com/graphql',
   method: 'POST' as const,
   supportedMethods: ['POST' as const],
-  onRun() {},
   onSetMethod(_method: HttpMethod) {},
 };
 
@@ -29,11 +20,14 @@ const renderTopBar = (ui: ReactElement) =>
   render(<Tooltip.Provider>{ui}</Tooltip.Provider>);
 
 describe('TopBarView', () => {
-  it('renders the Run button', () => {
-    render(<TopBarView {...DEFAULTS} />);
-    expect(
-      screen.getByRole('button', { name: /Run query/i }),
-    ).toBeInTheDocument();
+  it('renders the run button passed into the slot', () => {
+    render(
+      <TopBarView
+        {...DEFAULTS}
+        runButton={<button type="button">Run</button>}
+      />,
+    );
+    expect(screen.getByRole('button', { name: 'Run' })).toBeInTheDocument();
   });
 
   it('renders the GraphiQL wordmark', () => {
@@ -171,31 +165,6 @@ describe('TopBarView', () => {
     expect(onSetMethod).toHaveBeenCalledWith('POST');
   });
 
-  it('calls onRun when the Run button is clicked', async () => {
-    const user = userEvent.setup();
-    const onRun = vi.fn();
-    render(<TopBarView {...DEFAULTS} onRun={onRun} />);
-    await user.click(screen.getByRole('button', { name: /Run query/i }));
-    expect(onRun).toHaveBeenCalled();
-  });
-
-  it('disables the Run button while fetching', () => {
-    render(<TopBarView {...DEFAULTS} isFetching />);
-    expect(screen.getByRole('button', { name: /Run query/i })).toBeDisabled();
-  });
-
-  it('disables the Run button when a mutation is blocked over GET', () => {
-    renderTopBar(
-      <TopBarView
-        {...DEFAULTS}
-        method="GET"
-        supportedMethods={['GET', 'POST']}
-        runDisabledReason="Mutations can only be sent via POST"
-      />,
-    );
-    expect(screen.getByRole('button', { name: /Run query/i })).toBeDisabled();
-  });
-
   it('highlights the method toggle when a mutation is blocked over GET', () => {
     const { container } = renderTopBar(
       <TopBarView
@@ -210,25 +179,7 @@ describe('TopBarView', () => {
     ).not.toBeNull();
   });
 
-  it('wraps the disabled Run button in a focusable tooltip target when blocked', () => {
-    renderTopBar(
-      <TopBarView
-        {...DEFAULTS}
-        method="GET"
-        supportedMethods={['GET', 'POST']}
-        runDisabledReason="Mutations can only be sent via POST"
-      />,
-    );
-    // A native disabled button emits no events, so the tooltip needs a
-    // focusable wrapper to receive hover/focus and open.
-    const target = screen
-      .getByRole('button', { name: /Run query/i })
-      .closest('.graphiql-top-bar-run-tooltip-target');
-    expect(target).not.toBeNull();
-    expect(target).toHaveAttribute('tabindex', '0');
-  });
-
-  it('does not highlight the toggle, disable Run, or wrap it when not blocked', () => {
+  it('does not highlight the method toggle when not blocked', () => {
     const { container } = renderTopBar(
       <TopBarView
         {...DEFAULTS}
@@ -237,133 +188,12 @@ describe('TopBarView', () => {
       />,
     );
     expect(
-      screen.getByRole('button', { name: /Run query/i }),
-    ).not.toBeDisabled();
-    expect(
       container.querySelector('.graphiql-top-bar-method-toggle--attention'),
-    ).toBeNull();
-    expect(
-      container.querySelector('.graphiql-top-bar-run-tooltip-target'),
     ).toBeNull();
   });
 
   it('has role="banner" on the header element', () => {
     const { container } = render(<TopBarView {...DEFAULTS} />);
     expect(container.querySelector('header[role="banner"]')).not.toBeNull();
-  });
-
-  describe('operation picker', () => {
-    const TWO_OPS = opsOf('query Alpha { a }\nquery Beta { b }');
-
-    it('shows no caret with zero or one operation', () => {
-      const { rerender } = render(<TopBarView {...DEFAULTS} operations={[]} />);
-      expect(
-        screen.queryByRole('button', { name: 'Choose operation to run' }),
-      ).not.toBeInTheDocument();
-
-      rerender(
-        <TopBarView {...DEFAULTS} operations={opsOf('query Alpha { a }')} />,
-      );
-      expect(
-        screen.queryByRole('button', { name: 'Choose operation to run' }),
-      ).not.toBeInTheDocument();
-    });
-
-    it('shows no caret when an operation name is pinned via overrideOperationName', () => {
-      render(
-        <TopBarView
-          {...DEFAULTS}
-          operations={TWO_OPS}
-          overrideOperationName="Alpha"
-        />,
-      );
-      expect(
-        screen.queryByRole('button', { name: 'Choose operation to run' }),
-      ).not.toBeInTheDocument();
-    });
-
-    it('shows a caret with more than one operation', () => {
-      render(<TopBarView {...DEFAULTS} operations={TWO_OPS} />);
-      expect(
-        screen.getByRole('button', { name: 'Choose operation to run' }),
-      ).toBeInTheDocument();
-    });
-
-    it('lists every operation by name in the menu', async () => {
-      const user = userEvent.setup();
-      render(<TopBarView {...DEFAULTS} operations={TWO_OPS} />);
-      await user.click(
-        screen.getByRole('button', { name: 'Choose operation to run' }),
-      );
-      expect(
-        await screen.findByRole('menuitem', { name: 'Alpha' }),
-      ).toBeInTheDocument();
-      expect(
-        screen.getByRole('menuitem', { name: 'Beta' }),
-      ).toBeInTheDocument();
-    });
-
-    it('sets the operation name and runs when picking an operation', async () => {
-      const user = userEvent.setup();
-      const onRun = vi.fn();
-      const onSetOperationName = vi.fn();
-      render(
-        <TopBarView
-          {...DEFAULTS}
-          operations={TWO_OPS}
-          operationName="Alpha"
-          onRun={onRun}
-          onSetOperationName={onSetOperationName}
-        />,
-      );
-      await user.click(
-        screen.getByRole('button', { name: 'Choose operation to run' }),
-      );
-      await user.click(await screen.findByRole('menuitem', { name: 'Beta' }));
-      expect(onSetOperationName).toHaveBeenCalledWith('Beta');
-      expect(onRun).toHaveBeenCalled();
-    });
-
-    it('does not call setOperationName when picking the already-active operation', async () => {
-      const user = userEvent.setup();
-      const onRun = vi.fn();
-      const onSetOperationName = vi.fn();
-      render(
-        <TopBarView
-          {...DEFAULTS}
-          operations={TWO_OPS}
-          operationName="Alpha"
-          onRun={onRun}
-          onSetOperationName={onSetOperationName}
-        />,
-      );
-      await user.click(
-        screen.getByRole('button', { name: 'Choose operation to run' }),
-      );
-      await user.click(await screen.findByRole('menuitem', { name: 'Alpha' }));
-      expect(onSetOperationName).not.toHaveBeenCalled();
-      expect(onRun).toHaveBeenCalled();
-    });
-
-    it('disables a menu item whose operation is blocked for the current method', async () => {
-      const user = userEvent.setup();
-      const ops = opsOf('query Q { a }\nmutation M { b }');
-      renderTopBar(
-        <TopBarView
-          {...DEFAULTS}
-          method="GET"
-          supportedMethods={['GET', 'POST']}
-          transportMethod="GET"
-          operations={ops}
-        />,
-      );
-      await user.click(
-        screen.getByRole('button', { name: 'Choose operation to run' }),
-      );
-      const queryItem = await screen.findByRole('menuitem', { name: 'Q' });
-      const mutationItem = screen.getByRole('menuitem', { name: 'M' });
-      expect(queryItem).not.toHaveAttribute('data-disabled');
-      expect(mutationItem).toHaveAttribute('data-disabled');
-    });
   });
 });
